+++ apps/docs/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "./.nuxt/tsconfig.json",
+  // "extends": "./.nuxt/tsconfig.json",
   "ignoreConfigErrors": true
 }
