+++ packages/master/go.mod
@@ -1,0 +1,3 @@
+module master
+
+go 1.23.3
